ref: Split ServerRuntimeOptions into two w/ and w/o OTel
