+++ packages/bun/src/types.ts
@@ -1,12 +1,12 @@
 import type { BaseTransportOptions, ClientOptions, Options } from '@sentry/core';
-import type { ServerRuntimeOptions } from '@sentry/node-core';
+import type { OpenTelemetryServerRuntimeOptions } from '@sentry/node-core';
 
 /**
  * Base options for the Sentry Bun SDK.
- * Extends the common WinterTC options shared with Node.js and other server-side SDKs.
+ * Extends the common WinterTC options with OpenTelemetry support shared with Node.js and other server-side SDKs.
  */
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface BaseBunOptions extends ServerRuntimeOptions {}
+export interface BaseBunOptions extends OpenTelemetryServerRuntimeOptions {}
 
 /**
  * Configuration options for the Sentry Bun SDK
